+++ packages/form-render/src/form-items/date/index.tsx
@@ -66,8 +66,8 @@ export function renderDateItem<T = DataRecord>(options?: RenderDateItemOptions):
             v-model={data[form.key as keyof T]}
             disabled-date={disabledMethod}
             format={options?.labelFormat}
-            shortcuts={options?.shortcuts}
-            value-format={options?.valueFormat}></DatePicker>
+            value-format={options?.valueFormat}
+            {...{ shortcuts: options?.shortcuts }}></DatePicker>
         )
     }
   }