+++ packages/form-render/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gopowerteam/form-render
 
+## 0.0.64
+
+### Patch Changes
+
+- update datepicker shortruct
+
 ## 0.0.63
 
 ### Patch Changes