+++ packages/table-render/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gopowerteam/table-render
 
+## 0.0.113
+
+### Patch Changes
+
+- Updated dependencies
+  - @gopowerteam/form-render@0.0.64
+
 ## 0.0.112
 
 ### Patch Changes